Move html-filter into its own directory
