+++ ts/editor/index.ts
@@ -1,9 +1,10 @@
 /* Copyright: Ankitects Pty Ltd and contributors
  * License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html */
 
+import { filterHTML } from "html-filter/index";
+
 import { caretToEnd } from "./helpers";
 import { saveField } from "./changeTimer";
-import { filterHTML } from "./htmlFilter";
 import { updateButtonState, disableButtons } from "./toolbar";
 
 import { EditorField } from "./editorField";
