+++ ts/html-filter/helpers.ts
@@ -1,0 +1,7 @@
+export function isHTMLElement(elem: Element): elem is HTMLElement {
+    return elem instanceof HTMLElement;
+}
+
+export function isNightMode(): boolean {
+    return document.body.classList.contains("nightMode");
+}
