+++ ts/tsconfig.json
@@ -7,7 +7,8 @@
         "baseUrl": ".",
         "paths": {
             "anki/*": ["../bazel-bin/ts/lib/*"],
-            "sveltelib/*": ["../bazel-bin/ts/sveltelib/*"]
+            "sveltelib/*": ["../bazel-bin/ts/sveltelib/*"],
+            "html-filter/*": ["../bazel-bin/ts/html-filter/*"]
         },
         "importsNotUsedAsValues": "error",
         "outDir": "dist",
